MySQLi: various updates [15] - add mysqli_driver methods to versions
The OO-style methods were not included in the `versions.xml` file. This is a known issue based on the `FIXME` comment at the top of the file.

This commit adds the methods from the `mysqli_driver` class to the `versions.xml` file.

Version numbers based on the versions as set for the procedural functions. **This will need verification.**

Patch contributed by jrfnl.


git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@350072 c90b9560-bf6c-de11-be94-00142212c4b1