MySQLi: various updates [17] - add mysqli_warning methods to versions
The OO-style methods were not included in the `versions.xml` file. This is a known issue based on the `FIXME` comment at the top of the file.

This commit adds the methods from the `mysqli_warning` class to the `versions.xml` file.

Version numbers based on the versions as set for the procedural functions. **This will need verification.**

Includes removing the `FIXME` comments as AFAICS all methods have now been added.

Patch contributed by jrfnl.


git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@350073 c90b9560-bf6c-de11-be94-00142212c4b1