Removed misleading example from rowCount()
The example given was confusing and misleading. If the purpose of it is
to show how to get the count of matching records from the database then
the example can be much shorter. `query()` should not be checked with
`if` statement and there should not be 2 separate queries to the
database if the actual count is not needed. If the number of fetched
records is needed then you can simply use PHP's `count()` function.


git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@351083 c90b9560-bf6c-de11-be94-00142212c4b1